requirements.txt
Flask==2.1.2
future==0.18.2
gunicorn==20.1.0
h11==0.13.0
httplib2==0.14.0
ChatterBot==1.0.8
mysql-connector-python==8.0.29
